+++ tests/api-resources/trip.test.ts
@@ -8,12 +8,9 @@ const onebusawaySDK = new OnebusawaySDK({
   baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
 });
 
-describe('resource arrivalAndDepartureForStop', () => {
-  test('retrieve: only required params', async () => {
-    const responsePromise = onebusawaySDK.arrivalAndDepartureForStop.retrieve('1_75403', {
-      serviceDate: 0,
-      tripId: 'string',
-    });
+describe('resource trip', () => {
+  test('retrieve', async () => {
+    const responsePromise = onebusawaySDK.trip.retrieve('string');
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;
@@ -23,13 +20,10 @@ describe('resource arrivalAndDepartureForStop', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  test('retrieve: required and optional params', async () => {
-    const response = await onebusawaySDK.arrivalAndDepartureForStop.retrieve('1_75403', {
-      serviceDate: 0,
-      tripId: 'string',
-      stopSequence: 0,
-      time: 0,
-      vehicleId: 'string',
-    });
+  test('retrieve: request options instead of params are passed correctly', async () => {
+    // ensure the request options are being passed correctly by passing an invalid HTTP method in order to cause an error
+    await expect(onebusawaySDK.trip.retrieve('string', { path: '/_stainless_unknown_path' })).rejects.toThrow(
+      OnebusawaySDK.NotFoundError,
+    );
   });
 });